# specs/openapi/openapi-1a.yaml
openapi: 3.0.3
info:
  title: Incident Assistant API (Phase 1a)
  version: 1.0.0
  description: >
    Incidents + health. Normative: specs/phases/phase-1a-monolith-core/api-contract.md
servers:
  - url: http://localhost:8080
tags:
  - name: Incidents
  - name: Health
paths:
  /api/v1/incidents:
    get:
      tags: [Incidents]
      summary: List incidents (1a; items are manual-only; no source filter)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: ["createdAt,asc", "createdAt,desc"]
            default: "createdAt,desc"
      responses:
        "200":
          description: Paged IncidentSummary (manual incidents only in 1a)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedIncidents" }
        "400": { description: Bad query }
        "503": { description: Persistence unavailable }
    post:
      tags: [Incidents]
      summary: Create incident (DRAFT, MANUAL)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIncidentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "400":
          description: Validation (e.g. missing title or severity)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MinimalError" }
        "413": { description: Payload too large }
        "415": { description: Unsupported media type }
  /api/v1/incidents/{id}:
    get:
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: >
            Incident (Phase 1a wire shape). api-contract.md requires a strong ETag on this response
            for the Phase 1a gate; the Story 4 implementation does not emit ETag yet (see Story 5).
          headers:
            ETag:
              required: false
              description: >
                RFC 9110 opaque entity-tag (quoted decimal version). Implemented in Story 5; omitted in Story 4.
              schema: { type: string, example: '"7"' }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Incident" }
        "404": { description: Not found }
        "503": { description: Persistence unavailable }
    patch:
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200": { description: OK }
        "400": { description: Empty or invalid patch body }
        "404": { description: Not found }
        "409": { description: Not editable }
        "412": { description: Stale If-Match }
        "415": { description: Unsupported media type }
  /api/v1/incidents/{id}/transitions:
    post:
      tags: [Incidents]
      summary: State transition (If-Match required)
      parameters:
        - $ref: "#/components/parameters/IncidentId"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200": { description: OK }
        "400": { description: Invalid body, unknown to value, or reason too long }
        "404": { description: Not found }
        "409": { description: Illegal transition }
        "412": { description: Stale }
        "415": { description: Unsupported media type }
  /actuator/health:
    get:
      tags: [Health]
      responses:
        "200": { description: UP }
components:
  schemas:
    MinimalError:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
    CreateIncidentRequest:
      type: object
      required: [title, severity]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 32768
        severity:
          type: string
          enum: [SEV1, SEV2, SEV3, SEV4]
    IncidentSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status:
          type: string
          enum: [DRAFT, OPEN, CLOSED, CANCELLED]
        severity:
          type: string
          enum: [SEV1, SEV2, SEV3, SEV4]
        source:
          type: string
          enum: [MANUAL]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer, format: int64 }
      required: [id, title, status, severity, source, createdAt, updatedAt, version]
    Incident:
      allOf:
        - $ref: "#/components/schemas/IncidentSummary"
        - type: object
          properties:
            description: { type: string, nullable: true }
            transitionReason: { type: string, nullable: true }
    PagedIncidents:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/IncidentSummary" }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
      required: [items, page, size, totalElements, totalPages]
  parameters:
    IncidentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }